Change admin dashboard file structure
